refactor+bufgix
Make refactor
Fix bug with no scroll
Fix bug with no scroll-to-field
Fix bug no focus after submit
