+++ sample/iosSample/iosSample/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
