+++ sample/iosSample/iosSample/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
