+++ sample/iosSample/iosSample/Resources/Assets.xcassets/Contents.json
